calculate es scores
